:hammer: refactor for requested changes
